Make setup.py executable
Use env to choose python3, and make setup.py executable so that it
can be run in the common "./setup.py" way.  Remove the code that
removed --single-version-externally-managed from arguments; it
can be added again later if there is a specific reason for it.

Update .gitignore to future-proof for pyvenv use likely for use
with ./setup.py

Change-Id: Ice6db83d43d1cb8e04689b1f01eddafe54cb0517